+++ lib/resource/template/textfield/template/TextFieldTemplate.dart
@@ -130,7 +130,7 @@ class TextFieldTemplate   {
 
       //padding + hint + underline
       decoration: TextFieldTemplateBase.getDecorationInput(isShowBoarder, padding,
-          hint_text, hint_color, fontSize, isRemoveUnderline, prefixIcon, Colors.red,
+          hint_text, hint_color, fontSize, isRemoveUnderline, prefixIcon, null, Colors.red,
           autovalidateMode!, null ),
 
       //keyboard