+++ lib/resource/template/textfield/validator/ValidatorType.dart
@@ -1,3 +1,3 @@
 enum ValidatorType {
-  defaultValidator, email, password, phone, description, name, nameFull, moreThan2Char
+  defaultValidator, email, password, phone, description, name, nameFull, moreThan2Char, emailOrPhone
 }